Run preview-render-cli tool in external jvm process
The preview render CLI tool has a dependency to
the IntelliJ framework which uses ExecutorService
internally. Once you shut down ExecutorService,
you cannot restart the ExecutorService. This
makes the preview-render-cli non-reusable.

Since the Gradle worker API requires a worker
to be reusable for subsequent requests, we
cannot load the preview-render-cli tool's
classes in a worker daemon.

Fixes: 340362066
Test: //tools/base/build-system/integration-test/application:screenshot-tests
Change-Id: I58f74ee340445307368c9f3305e166805a2a35e5

Former-commit-id: 186e0828b47830bfdefcd64fc33c620d78b7d31a