Always pass binary resources to resource shrinker task
ShrinkResourcesNewShrinkerTask currently takes the following resources
as inputs:
  a) LINKED_RESOURCES_BINARY_FORMAT if multi-APK is enabled, or
  b) LINKED_RESOURCES_FOR_BUNDLE_PROTO_FORMAT if multi-APK is not
     enabled (the common case -- building a regular APK)

The downside of (b) is that it requires LinkAndroidResForBundleTask to
run to produce resources in proto format, which requires running `aapt2
link` the second time (the first `appt2 link` run is done in
LinkApplicationAndroidResourcesTask, which is always required when
building an APK).

To ensure `aapt2 link` runs only once, this change updates
ShrinkResourcesNewShrinkerTask such that it always takes as input
LINKED_RESOURCES_BINARY_FORMAT. The task will run `appt2 convert` to
convert binary resources to proto format first before passing proto
resources to the resource shrinker, but `aapt2 convert` is much cheaper
than `appt2 link`.

Bug: 326033134
Test: Updated ResourceShrinkerTest
Change-Id: I1d3d649427a920c45ffecff6d1859138aab946f4

Former-commit-id: b8ecd4228fe1dd5c75ac5d3c54298828d6ce7abc