Consolidate AVD .ini file keys
Instead of having most keys defined in AvdManager, and a few in
AvdWizardUtils, put every key in one of MetadataKey, ConfigKey, or
UserSettingsKey, depending on which config file it is for.

This makes the huge AvdManager class easier to read and avoids the need
for the AVD_INI_ prefix.

Also, improve the documentation of the keys.

Bug: 339875083
Test: existing
Change-Id: Ic96f76b4a038cda9da3d503dd0f45ab3e70684e7

Former-commit-id: 83d420b4d950645063036b58b78e27103a33cb09