349491177: Don't enforce control flow on wakelocks stored in fields
If the wakelock is a field, it's quite possible/likely that the release
is in a different method, even if we have a local release. Until this
CL, the code assumed that if we see both an acquire and a release in the
same method, this is a local use of the wakelock which was always
intended to be cleaned up -- so we perform control flow analysis and
warn if there's any escape from the method not going through the release
call.

But as b/349491177 shows, another possible scenario is releasing locks
if some operation failed but otherwise leaving it for another method.

Test: Unit test included
Fixes: 349491177
Change-Id: I215f8945b8d5fd022132e610974cc6787cc7d606

Former-commit-id: 35541f44b0fdce7b58f64516891c87c4d96aaa25