347356457: Fix false positive string concatenation SetTextI18n warning
We have an old lint check which warns you if you're performing string
concatenations and passing the result to one of the
view text methods.  Recently this started warning about concatenations
for simple strings.

The root cause is probably the recent UAST change to model all Kotlin
strings as UPolyadicExpressions with the plus operator, which the code
interpreted as concatenations. This CL fixes this issue to only warn if
there are at least two operands.

(It also adds a positive test for Kotlin to make sure we do warn for
String interpolations passed to view methods.)

Test: Unit test included
Fixes: 347356457
Change-Id: Ib8f1ce0224b7cf0c1f94d37b06dd2cfd5dcddef2

Former-commit-id: e51a3d4cb8b19354fc9399848517246d1b3a29df