[SDK Index] Do not show URL for non SDK Index libraries
The provided link is to prevent older versions of Studio to get an empty
link and thus there is no point in generating a link if it does not
provide information. With this change first we first check if the link
does not come from the SDK Index and and otherwise returns it.

This change is based on commit 18dfce70f48419ebc502a0b9c7b275a4a6830b21 [formerly 13e6fad89257ba96485e93373739560cad12037f]
with some differences in GradleDetectorTest to make sure SDK from
different sources are used.

Bug: 339094763
Test: GooglePlaySdkIndexTest, GradleDetectorTest
Change-Id: Ic969663cd6e724cf2c6f0549d05c75bcfe5a6bad

Former-commit-id: 039c4d5924428233ea2c9f6253ac02582a6200a6