Fix bugs in the constant evaluator
The constant evaluator has code paths for both handling UElements and
PsiElements. The UElement support is more feature rich; it has direct
support for considering the size of constant arrays etc.

This all started when I discovered some surprising results in lint's
extracted issue documentation; test files containing the following
string:
    <resources>
    </resources>
were instead recorded as
    <resources>null</resources>

Where is the null coming from? This happened because lint's issue
extraction analyzes unit test source files; normally test contents are
included directly as literal strings, but in some cases they are stored
into a local property or field such that they can be shared between
tests.

And when the constant evaluator evaluates a reference, it of course
resolves the reference -- so now it's operating in PSI space -- and then
it uses the PSI part of the constant evaluator.

And the first bug here is that sometimes, Kotlin template strings (when
you have an escape, like \n), there's a KtEscapeStringTemplateEntry
element instead which needs to be handled.

So I updated the test for this, and realized that our constant evaluator
tests weren't actually testing the reference-indirection, so I updated
the tests to do that, and then I discovered various other failures.
These are shown by the additional unit tests added in this CL.

Test: Unit tests included
Bug: Not filed
Change-Id: Idaadb88bc45b16865e6a20ead5def9b070ffb973

Former-commit-id: 6ca9fc2a24a345a5156eb8222a6d65fe8c7ee1f4