Fix AndroidManifest.xml in rsSupportMode project
Also rename the "image" directory to "image2" to match the package name.

Without this change, running lint on the rsSupportMode project with K2
UAST results in a MissingClass lint issue because the fully qualified
class name in the manifest
("com.android.rs.image.ImageProcessingActivity2") doesn't exist.

When running lint without K2 UAST, there's no such lint issue because
the fully qualified class name matches the relative path. This is a bug,
but it's beyond the scope of this fix.

Bug: 353683286
Test: existing
Change-Id: Ia4b93ff9f10d937af18cf7b09b47f331d6841d63

Former-commit-id: d613cfce21074270aa098a0c84902a969265f973