Improve lint test infrastructure class name scanning
The code to automatically infer the package and class name from Java and
Kotlin files, used in the test suite (such that you don't have to
declare test file names directly) uses regular expressions to find the
package and class/object names.  This is using regular expressions, not
full blown parsing.

However, a file can contain a comment before the class definition, which
confuses the regular expressions. Therefore, we strip comments from the
source before searching for the pattern. However, with annotations that
take string parameters, we can also have text in those strings which
confuses the pattern matching. Therefore, this CL updates the comment
stripping to both strip annotations *and* remove comments before
searching for the class name. (We already have code to do this that we
can reuse from the quickfix machinery).

Test: Existing test and new test for the specific failure scenario
Bug: Not filed
Change-Id: Ib5dbfdc3f686ef1e27619a59ccac983cae874755

Former-commit-id: 395ea73daa75f0218d1f475712dacddbe812fb61