Provide dependency model with a flat list on AGP
We intended to use the graph in the IDE, but it comes with a large
performance cost and is currently unused, so switching to a flat list
of dependencies instead for each variant.

This is expected to improve the performance of Android model building
by not asking for the resolution result (i.e. the graph), therefore
avoiding additional work.

However, there is currently a regression with Gradle 8.7 that introduces
some lock contention, which means the improvements will not be as
visible until we land a version with a fix (I've confirmed this is the
case by "fixing" the bug locally on Gradle)

Test: Covered by snapshot tests
Bug: 354906596
Change-Id: I460dceb73bbd9dbdb8171064a456c92ad2df7a86

Former-commit-id: 6e610eff22ff7f3bad35c6e8a9644817b9ada30c