Use Isolated project view to access rootProject dir
IsolatedProject interface has much less functions than a Project
interface. Surface a message to users when no profile outputs
directory is configured because we there could be a potential issue
when users use a different buildDir than the default(/build), so
that then they see we create a separate build directory, they are
informed of the cause and provided a way to fix it.

Bug: 314817109
Test: ProjectIsolationWithAnalyticsTest
Change-Id: I069f6dabe67f87064622e846d9a62019603df0ad

Former-commit-id: 04c2df5c00b9abd6c54ee236e8a77f1aaf30f029