[privacy sandbox] check if there are SDKs before verifying device support
There is an issue with ddmlib on Linux ARM that causes a timeout when
running shell commands. Since the
ConnectedDevice.getSupportsPrivacySandbox is partly based on running a
shell command, this change prevents the check unless there are SDKs
present. The bug will still remain if there are SDKs, but this will
mitagate for most installs.

Bug: 354862210
Test: N/a
Change-Id: Iac2ce010a8610bd1c88a9ded99caef839f074c9e

Former-commit-id: 7bce2ba72b96a6bcbf34ad1b9804cf4c017d60e5