Make variable naming more uniform
AVDs consist of various files: the metadata .ini file, the AVD data
folder, and various files within the AVD data folder, like config.ini.
Many of the variable names used do not make it clear which of the above
they refer to.

The following names are used:
avdName: the basename of the metadata .ini file
metadataIniFile: the Path to the metadata .ini file
avdFolder: the AVD's data folder (typically ending in .avd)

Bug: 339875083
Test: AvdManagerTest
Change-Id: Id7e1d69f285b844d03284963ba78cf6e660bffba

Former-commit-id: 45009edeb93fc0b447fadeae4235d579571e9c7b