Postprocessing block not applied in optimization
When initWith is used for a build type, the postprocessing block's
properties are not being applied to the build type in the
OptimizationDslInfoImpl because the postProcessingBlockUsed property is
not being propagated from the initial build type. Propagating the
property correctly in the _initWith method in BuildType causes the files
to be correctly gathereed in the optimization object.

Bug: 307784512
Test: ProguardFilesTest
Change-Id: I169995202842605e870180244dba555b9c47d93c

Former-commit-id: 08356be7b624ad785f01b390554b80b14838229f