Add tests for remaining presubmit functions
This CL adds tests for presubmit.find_test_targets and presubmit.generate_runs_per_test_flags. The tests cover the following cases:

* A change with a default Presubmit-Test tag
* A change with a Presubmit-Test tag that specifies a different build target
* A change with a Presubmit-Test tag that specifies explicit targets
* A change with Presubmit-Runs-Per-Test that specifies no build target, the current build target, and a different build target.

The tests also verify that the correct flags are passed to the bazel query command.

This change also fixes a couple of minor bugs found by the tests.

Bug: 353307795
Change-Id: I35b90417a9fc71e81ee84bd7f5cd12d2420b7f15

Former-commit-id: 359dae62ce0475da40d140e4bf047fbaf635ebe1