Consistent settings file generation in groovy and kts format
Settings.gradle and settings.gradle.kts are both used in our test
projects today. Previously, we only consider settings.gradle as the
settings file of the project and our altering logic only applies to it,
instead of the settings.gradle.kts. This would prevent us from using
versionCatalog in projects with kts files.

Bug: n/a
Test: existing
Change-Id: I572419394cdc550445926c1a4684bb47b9a5df00

Former-commit-id: 551580903df7d70684ceceb849302c6f0815de1d