Improvements to the bytecode rewriter and testing
First, this CL improves the tests for the bytecode rewriter.  Before
this CL, the tests would check the bytecode diffs for known 3p lint
utility methods, confirming they're as expected -- based on manual
testing in an Android project.

Instead, this CL writes simpler utility methods for all the same
senarios -- compiled against the same older analysis API.  After
performing bytecode verification, it then actually loads and *runs* the
bytecode and confirms that the migrated code correctly runs. This means
that in the future, when the analysis API changes again, the bytecode
diffs will continue to be the same, but we'll discover if the code no
longer runs correctly.

This CL also updates the bytecode rewriter itself in two ways: First,
there was a change to the AndroidX LintUtils class which was special
cased, and now needs to be updated to deal with default arguments. And
second, when verifying execution of the migrated code I discovered that
the allSuperTypes method was returning a sequence rather than a list, so
updated the translation.

Test: Many included unit tests
Bug: 360840930
Change-Id: I8c5a65ad97c63786c803b4178dd11c009bbd8dd8

Former-commit-id: d0dbf29ae8a7e7fe500a4f67592fd2ef54168f57