Add BooleanOption to run resource shrinking in R8
This change introduces the following Boolean Gradle property:

    android.r8.integratedResourceShrinking

When enabled, the R8 task will perform resource shrinking in addition to
code shrinking. When disabled, resource shrinking will be performed in a
separate task after the R8 task has run (previous behavior before this
change).

Note: If resource shrinking is not enabled, this flag has no effect.

The flag is currently disabled by default. We will enable it by default
in the next change. In AGP 9.0, it will always be enabled with no option
to disable.

Test: Updated ResourceShrinkerTest
Bug: 326033134
Change-Id: I19c72e8c31c3c0ea66f662f0e646dd7761fa36d6

Former-commit-id: c0b54dd20d723bb7365842b5add6b742a0543acd