Ignore ERROR_DEVICE_MISSING condition
This means that the device profile that was used to create the AVD is no
longer present. This can happen if the user deletes a custom device
profile, or if the user had created an AVD with a newer version of
Studio that has built-in device profiles that are not present in the
current version.

This doesn't matter; we don't need the device profile after the AVD is
created. While we do scan for changes in the device profile and update
the AVDs created from it accordingly, this is just a convenience. If the
device profile is completely missing, we can just proceed as is.

Fixes: 315168101
Test: n/a
Change-Id: I72d749a3a5c76e4bd244b88cc134a11725c3133d

Former-commit-id: 52a2aa0ba6e2566bbaf90ac0a1b0ba4b4923fc63