Introduce ReplacedByIncubating annotation.
When replacing a stable API with a new incubating one, it is not
possible to immediately deprecate the old API as it leaves the users
havign to choose between a Deprecated or an Incubating API.

Instead of deprecating the old API, annotate it with @ReplacedByIncubating
annotation so we can fail a test when the incubating API graduates to
stable and the old API does not become Deprecated.

This will also indicate to users that this API will become deprecated
soon.

Test: Added a unit test to check that @ReplacedByIncubating annotated
element are replaced by an incubating API, which should fail as soon
as the API graduates.
Bug: N/A

Change-Id: I3aab1c589021ff67230f416f0cde8be500badfee

Former-commit-id: e13ee2051f09ea8d57eecb99d87115630aa2b60b