Reorder AVD creation to avoid unnecessary deletion
Exceptions that occur after we've written anything to disk cause the
whole AVD to be deleted, presumably because we don't know what state
it's in. Do as much work as we can before writing anything to disk to
minimize the chance that this occurs.

Inline createAvdConfigFile, which just calls setImagePathProperties.

Bug: 368538172
Test: AvdManagerTest
Change-Id: I7c5a3bd57291c87738bca938ac00a77107a117cd

Former-commit-id: 560b187ff96dcb0f3456f9f2c29c34807c544d5f