Compute fixpoint of monotones over finite functions on `Lattice`
This change provides a `Monotone` interface (which is basically
`(X → Y) → (X → Y)`, subject to a few more invariants), and implements
a baseline algorithm for computing its fixpoint.

This allows an analysis to be implemented as a straightforward
open-recursive function, agnostic of how it's run (e.g. in optimizal
order, in parallel, etc.), and how it's used (e.g. whole-module, on a
few methods, or incrementally, etc.).

Bug: n/a
Test: new
Change-Id: I4c066bcb6ed604ad2c30775acf3cef43cdd4b5fa

Former-commit-id: 5e1bc09c44e5c69ac39fbddf5bb1d4f14dd1f46e