Allow pseudolocales to be present even when localeFilters are present
Adds the pseudolocales under the hood to the locale filters/resource
configurations, when either of the two have been specified. Otherwise,
specifying them will filter out the pseudolocales when enabled. If users
*want* to filter out the pseudolocales, they should just turn off the
feature. This makes it so that they do not have to add `en-rXA` and
`ar-rXB` manually to the locale filters when they expect it in their
app.

Bug: 373506504
Test: updated
Change-Id: Ice77cc98434e323e99a947306d26cd1128201fa6

Former-commit-id: b1ee87cc92da1c533efe4fb4da1e1b3b1e9a6006