Lint: remove testSetExitCode()
SecurityManager is deprecated in the JDK, and soon it will no longer
be possible to intercept System.exit() calls.

Removing this test case seems OK to me. We are unlikely to have a
regression here, especially given that the System.exit() call is in a
single conspicuous place within Main.main().

Bug: 372751404
Test: this change removes a test
Change-Id: I7e3775e83d6cef4c135d1096acd0d80c9568274e

Former-commit-id: c32b7a33c73af86bcc237dcf0195345a7eee0250