New Builder interfaces to create a project
This is going to be part of the fixture (JUnit Rule) replacing
GradleTestProject.

This is a new set of builders that replace the TestProjectBuilder.
One key goal is to have direct access to the AGP extensions instead
of duplicating them in separate interfaces.

The other goal is to write projects using more modern Gradle
(particularly around plugin definition, buildScript and settings
content) and this was a lot simpler to create a new interface hierarchy
to allow for a smoother migration.

For now this can create the build files but not the content of the
project. This will come next.

Bug: N/A
Test: These are test fixtures but they also are unit tested.
Change-Id: Iab8102bb594b6313b20809eb5e3ff5061fe3b3be

Former-commit-id: ca6f729da948d97049fa13908979ceecef09579d